new version

+++ server/src/categories/category.service.ts
@@ -19,6 +19,9 @@ export class CategoryService {
       take: limit,
       skip,
       relations: ['products'],
+      order: {
+        order: 'ASC',
+      },
     });
     return paginateResponse(data, page, limit);
   }
@@ -41,23 +44,57 @@ export class CategoryService {
     await this.categoryRepository.update(id, { ...data });
   }
 
+  async changeOrder({ dndItems }) {
+    for (let i = 0; i < dndItems.length; i++) {
+      const thisItem = dndItems[i];
+      // const item = await this.categoryRepository.findOne({ id: thisItem.id });
+      await this.categoryRepository.update(thisItem.id, {
+        ...thisItem.id,
+        order: i + 1,
+      });
+    }
+    return true;
+    // const fromItem = await this.categoryRepository.findOne({ id: from.id });
+    // const toItem = await this.categoryRepository.findOne({ id: to.id });
+    // console.log(fromItem, toItem, oldIndex, newIndex);
+    // if (fromItem && toItem) {
+    //   await this.categoryRepository.update(fromItem.id, {
+    //     ...fromItem,
+    //     order: newIndex + 1,
+    //   });
+    //   await this.categoryRepository.update(toItem.id, {
+    //     ...toItem,
+    //     order: oldIndex + 1,
+    //   });
+    //   return true;
+    // }
+    // const orderedItemsIds = orderedItems.map((item) => item.id);
+    // console.log(orderedItemsIds);
+    // const currentCategories = await this.categoryRepository.find();
+    // await this.categoryRepository.update();
+    // return currentCategories;
+  }
+
   async delete(id) {
     const categoryToDelete = await this.categoryRepository.findOne(id);
     if (categoryToDelete) {
       await this.categoryRepository.delete(id);
       const categoryToDeleteImage = categoryToDelete.cover_image;
-      const pathToDelete = join(
-        process.cwd(),
-        'public',
-        'categories',
-        categoryToDeleteImage,
-      );
-      if (fs.existsSync(pathToDelete)) {
-        try {
-          fs.unlinkSync(pathToDelete);
-          //file removed
-        } catch (err) {
-          console.error(err);
+
+      if (categoryToDeleteImage) {
+        const pathToDelete = join(
+          process.cwd(),
+          'public',
+          'categories',
+          categoryToDeleteImage,
+        );
+        if (fs.existsSync(pathToDelete)) {
+          try {
+            fs.unlinkSync(pathToDelete);
+            //file removed
+          } catch (err) {
+            console.error(err);
+          }
         }
       }
     }
